Log file and issn A log file with dictionary keys with the format: JournalName.Vol.Pag to identify each processed entry have been implemented
Whenever the log file is presente only the new entries in
the input csv file is processed.

In addition a first implementation of issn generation is now included.
The issn is obtained after ordering the output of the publindex
spreadsheet in journal category in descending order.

A new dictionary with the alias used to name the Journals through
the csv file was started.